インスペクター非表示属性・カメラズーム等エディタ拡張
SerializedClassにHideInInspector属性を追加し、インスペクターでのフィールド表示制御を実装。Camera/EditorCameraのズーム・移動機能を強化し、Direct3D描画も対応。階層ウィンドウの非アクティブ表示やObject::Destroyの子再帰削除、型安全性向上、その他細かなUI・機能改善を含む。

+++ InvaderGame/Animator.cpp
@@ -27,7 +27,7 @@ void Animator::Play(std::string animationName)
 	auto itr = std::find(result.begin(), result.end(), animationName);
 	__int64 index = std::distance(result.begin(), itr);
 
-	if (index >= _animationVector.size())
+	if (index >= (int)_animationVector.size())
 	{
 		return;
 	}
